a little bit of reorganization
