+++ app/src/main/java/com/elliottsoftware/calftracker/presentation/components/newCalf/NewCalfViewModel.kt
@@ -28,7 +28,7 @@ data class NewCalfUIState(
     val birthWeight:String="",
     val sex:String="Bull",
     val birthDate:Date = Date(),
-    val calfSaved:Response<Boolean> = Response.Success(false),
+    val calfSaved:Response<Boolean> = Response.Failure(Exception()),
     val loggedUserOut:Boolean = false,
